pyproject.toml
[project]
name = "workflow-automation-aiops"
version = "0.1.0"
description = "AI-assisted assessment evaluation system with human-in-the-loop workflow"
authors = [
    {name = "Saurabh Bhardwaj", email = "saurabh@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ai", "assessment", "evaluation", "langchain", "langgraph", "openai", "streamlit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic-settings==2.6.1",
    "pypdf2==3.0.1",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "reportlab==4.2.5",
    "streamlit==1.39.0",
    "typing-extensions==4.12.2",
]
